Remove folly headers from public includes
The folly headers were previously included in the public include files.
This required that consumers of the library would need to at the very
least have the folly headers installed on their system.

This removes the folly headers from the public include directories,
forward declaring them or explicitly removing non-public headers from
the public include folder.
